requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
ultralytics==8.1.18
python-multipart==0.0.9
pillow==10.2.0